Cleaned up Hyperion::fake code
