+++ lib/hyperion/aux/hash_ext.rb
@@ -1,0 +1,5 @@
+class Hash
+  def subhash?(hash)
+    each_pair.all?{|k, v| hash[k] == v}
+  end
+end
