+++ lib/hyperion_test/fake_server/types.rb
@@ -1,0 +1,7 @@
+class Hyperion
+  class FakeServer
+    MimicRoute = ImmutableStruct.new(:method, :path)
+    Rule = ImmutableStruct.new(:mimic_route, :headers, :handler, :rest_route)
+    Request = ImmutableStruct.new(:body)
+  end
+end
